Bot/requirements.txt
colorama==0.4.4
ConfigArgParse==1.5.3
PyYAML==6.0
uiautomator2==2.16.19
urllib3==1.26.7
emoji==1.6.1
langdetect==1.0.9
atomicwrites==1.4.0
spintax==1.0.4